Crate bevy_utils

General utilities for first-party Bevy engine crates.

Re-exports

pub use hashbrown;

pub use petgraph;

pub use thiserror;

pub use tracing;

Modules

futures

Utilities for working with Futures.

label

Traits used by label implementations

prelude

synccell

A reimplementation of the currently unstable std::sync::Exclusive

syncunsafecell

A reimplementation of the currently unstable std::cell::SyncUnsafeCell

Macros

all_tuples

define_boxed_label

Macro to define a new label trait

define_label

Macro to define a new label trait

detailed_trace

Like tracing::trace, but conditional on cargo feature detailed_trace.

Structs

AHasher

A Hasher for hashing an arbitrary stream of bytes.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

FixedState

A hasher builder that will create a fixed hasher.

FloatOrd

A wrapper for floats that implements Ord, Eq, and Hash traits.

Hashed

A pre-hashed value of a specific type. Pre-hashing enables memoization of hashes that are expensive to compute. It also enables faster PartialEq comparisons by short circuiting on hash equality. See PassHash and PassHasher for a “pass through” BuildHasher and Hasher implementation designed to work with Hashed See PreHashMap for a hashmap pre-configured to use Hashed keys.

OnDrop

A type which calls a function when dropped. This can be used to ensure that cleanup code is run even in case of a panic.

PassHash

A BuildHasher that results in a PassHasher.

PassHasher

A no-op hash that only works on u64s. Will panic if attempting to hash a type containing non-u64 fields.

Uuid

A Universally Unique Identifier (UUID).

Traits

PreHashMapExt

Extension methods intended to add functionality to PreHashMap.

Functions

default

An ergonomic abbreviation for Default::default() to make initializing structs easier. This is especially helpful when combined with “struct update syntax”.

get_short_name

Shortens a type name to remove all module paths.

Type Definitions

BoxedFuture

An owned and dynamically typed Future used when you can’t statically type your result or need to add some indirection.

Entry

A shortcut alias for hashbrown::hash_map::Entry.

HashMap

A HashMap implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

HashSet

A HashSet implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

Instant

PreHashMap

A HashMap pre-configured to use Hashed keys and PassHash passthrough hashing.

StableHashMap

A stable hash map implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

StableHashSet

A stable hash set implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.