Module bevy_utils::nonmax

Structs

NonMaxI8

An integer that is known not to equal its maximum value.

NonMaxI16

An integer that is known not to equal its maximum value.

NonMaxI32

An integer that is known not to equal its maximum value.

NonMaxI64

An integer that is known not to equal its maximum value.

NonMaxI128

An integer that is known not to equal its maximum value.

NonMaxIsize

An integer that is known not to equal its maximum value.

NonMaxU8

An integer that is known not to equal its maximum value.

NonMaxU16

An integer that is known not to equal its maximum value.

NonMaxU32

An integer that is known not to equal its maximum value.

NonMaxU64

An integer that is known not to equal its maximum value.

NonMaxU128

An integer that is known not to equal its maximum value.

NonMaxUsize

An integer that is known not to equal its maximum value.

ParseIntError

An error type returned when an integer cannot be parsed (mimics std::num::ParseIntError)

TryFromIntError

An error type returned when a checked integral type conversion fails (mimics std::num::TryFromIntError)