Struct bevy_utils::nonmax::NonMaxUsize

#[repr(transparent)]
pub struct NonMaxUsize(/* private fields */);

An integer that is known not to equal its maximum value.

Implementations

impl NonMaxUsize

pub const fn new(value: usize) -> Option<NonMaxUsize>

Creates a new non-max if the given value is not the maximum value.

pub const unsafe fn new_unchecked(value: usize) -> NonMaxUsize

Creates a new non-max without checking the value.

Safety

The value must not equal the maximum representable value for the primitive type.

pub const fn get(&self) -> usize

Returns the value as a primitive type.

pub const ZERO: NonMaxUsize = _

Gets non-max with the value zero (0)

pub const ONE: NonMaxUsize = _

Gets non-max with the value one (1)

pub const MAX: NonMaxUsize = _

Gets non-max with maximum possible value (which is maximum of the underlying primitive minus one)