bevy_utils

Module prelude

The utilities prelude.

This includes the most common types in this crate, re-exported for your convenience.

Re-exports

pub use crate::default;

pub use crate::default;

Structs

DebugName

Wrapper to help debugging ECS issues. This is used to display the names of systems, components, …

ShortName

Lazily shortens a type name to remove all module paths.