Crate bevy_utils

Re-exports

pub use tracing;

Structs

AHasher
A Hasher for hashing an arbitrary stream of bytes.

Duration
A Duration type to represent a span of time, typically used for system timeouts.

FixedState
A hasher builder that will create a fixed hasher.

Uuid
A Universally Unique Identifier (UUID).

Type Definitions

BoxedFuture

HashMap
A std hash map implementing AHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

HashSet
A std hash set implementing AHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

Instant

StableHashMap
A stable std hash map implementing AHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

StableHashSet
A stable std hash set implementing AHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.